Swap values extracted to bounding box (ymin,ymax)
This original assigment of values resulted in ymax < ymin

The GTSDB README.txt states that the bounding box coordinates
arise on each line of raw_metadata.txt in the order
xmin;ymax;xmax;ymin. However, the bounding box coordinates are
with respect to an origin in the top left corner of the image and
so this assignment of values to variables leads to ymax < ymin.
Rather, it is better to extract the values in the form
xmin;ymin;xmax;ymax.  This is expected by the tensorflow object
detection api.